backend_py/phone-number-detection/requirements.txt
Flask>=2.2
Flask-CORS>=4.0.0
httpx>=0.24.0
phonenumbers>=8.13.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
